#27 Move the config enums to their own file
