+++ frontend/src/components/tabs/FloorConfigTab.tsx
@@ -1,7 +1,7 @@
 import HelpIcon from '../forms/HelpIcon';
+import ModConfig from '../../types/Configuration';
 import ContentContainer from '../layout/ContentContainer';
 import LabelledDropzone from '../dropzone/LabelledDropzone';
-import ModConfig, { Floor, ModOptions, Room } from '../../types/Configuration';
 
 import { Flex } from '@chakra-ui/react';
 import { useDispatch } from 'react-redux';
@@ -11,6 +11,7 @@ import { capitalise } from '../../utility/Utils';
 import { setFloorSkills } from '../../redux/slices/mod';
 import { hasOptionSet } from '../../utility/ConfigHelpers';
 import { saveSelectedMod } from '../../redux/slices/saving';
+import { ModOptions, Floor, Room } from '../../types/enums/ConfigEnums';
 
 interface FloorConfigTabProps {
     selectedMod: ModConfig;
