+++ frontend/src/utility/ConfigHelpers.ts
@@ -1,13 +1,7 @@
-import ModConfig, {
-    DEFAULT_CONFIG,
-    Floor,
-    FloorSkills,
-    ModOptions,
-    GlobalOptions,
-    Options,
-    Room,
-} from '../types/Configuration';
+import ModConfig, { DEFAULT_CONFIG, FloorSkills, Options } from '../types/Configuration';
+
 import { WeightedSkill } from '../types/Skill';
+import { ModOptions, GlobalOptions, Floor, Room } from '../types/enums/ConfigEnums';
 
 export function hasOptionSet(option: Options, flag: Options) {
     return flag === (option & flag);
